refactor: Moved app to Presentation
