+++ entry_point/app.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-use \Pryaniki\App\App;
+use Pryaniki\App\Presentation\Console\ConsoleEmailValidatorRunner;
 
 require '../app/vendor/autoload.php';
 
-$app = new App();
+$app = new ConsoleEmailValidatorRunner();
 
 $result = $app->run();
 echo $result;
